move ptt.
